separate eep-data observer from api-creation
