+++ web-server/src/server/clientio/socket-service.ts
@@ -19,9 +19,9 @@ export default class SocketService {
         addSocketEvents(socket);
       }
 
-      socket.on(RoomEvent.JoinRoom, (rooms: { room: string }) => {
-        socket.join(rooms.room);
-        if (this.debug) console.log(socket.id + ' joined rooms: "' + rooms.room + '" ');
+      socket.on(RoomEvent.JoinRoom, (room: { room: string }) => {
+        socket.join(room.room);
+        if (this.debug) console.log(socket.id + ' joined rooms: "' + room.room + '" ');
       });
 
       socket.on(RoomEvent.LeaveRoom, (room: { room: string }) => {
