+++ web-server/src/server/json/observers/state-observer.ts
@@ -1,0 +1,8 @@
+import { Socket } from 'socket.io';
+import JsonDataStore from '../json-data-reducer';
+
+export default interface StateObserver {
+  onStateChange: (store: Readonly<JsonDataStore>) => void;
+  onJoinRoom: (socket: Socket, room: string) => void;
+  onLeaveRoom: (socket: Socket, room: string) => void;
+}
